packages/api/pyproject.toml
[project]
name = "agentaudit-api"
version = "0.1.0"
description = "AgenticAudit API server"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlmodel>=0.0.22",
    "psycopg2-binary>=2.9.0",
    "pydantic-settings>=2.6.0",
    "alembic>=1.14.0",
    "python-ulid>=3.0.0",
    "jinja2>=3.1.0",
    "httpx>=0.28.0",
    "fpdf2>=2.8.0",
    "pyyaml>=6.0",
]

[project.scripts]
agentaudit-purge = "agentaudit_api.services.retention:_cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agentaudit_api"]
